Create common helpers and refactor charts
